Support requests greater than 2 GiB
git-svn-id: https://mockwebserver.googlecode.com/svn/trunk@44 cf848351-439f-e86a-257f-67fa721851d5

+++ src/main/java/com/google/mockwebserver/MockWebServer.java
@@ -556,7 +556,7 @@ public final class MockWebServer {
      * An output stream that drops data after bodyLimit bytes.
      */
     private class TruncatingOutputStream extends ByteArrayOutputStream {
-        private int numBytesReceived = 0;
+        private long numBytesReceived = 0;
         @Override public void write(byte[] buffer, int offset, int len) {
             numBytesReceived += len;
             super.write(buffer, offset, Math.min(len, bodyLimit - count));
